[refactor] 폴더구조 수정
- 폴더구조 수정
- 지나치게 나누어진 폴더 병합
- 한 depth에 너무 많은 파일이 있는 폴더 쪼개기
